Fix Deprecation Warning for Server Discovery and Monitoring
When using the module we are getting a DeprecationWarning that current Server Discovery and Monitoring engine is deprecated and will be removed in a future version. To disable the warning, we need to pass the option { useUnifiedTopology: true } on the constructor, howver, since the module uses Connection.openUri we need to update mongoose to a newer version as per the link on mongoose github.
- https://github.com/Automattic/mongoose/issues/8146